aix: Fix Node test test-tls-ocsp-callback.js
In debugging test-tls-ocsp-callback.js for the 0.12.0 Node on
AIX we discovered that there was code missing in
uv__platform_invalidate_fd() in aix.c which removes the file
descriptor from the pollset when needed.  We based the
impelementation in aix.c on the unix version in linux-core.cc.
We can see that this issue was addressed for linux in commit
780d8ad8e58c675dc60f5b73ea22a203760da2f2 to address joyent/libuv
issue #1099. This change introduces the same fix for AIX.
This change does not introduce any new failures when running
the tests on AIX. It also does not fix any existing failures,
but it does allow the node test test-tls-ocsp-callback.js to pass.

    modified:   src/unix/aix.c