thread: get affinity support for Windows.
Use two calls to SetThreadAffinityMask() to get the thread affinity, as
suggested by @petkaantonov. Get and use the process affinity mask for
the first call to hopefully reduce the probability of the thread being
forced to migrate in between the first call and the second call that
resets the old mask. Also, added validation against the process affinity
mask in uv_thread_setaffinity().