unix: removed double-checked read from uv_async
This provides only a negligible increase in performance, but can cause
incorrect results. The ACCESS_ONCE() macro only provides a weak compiler
barrier on GCC. The problem is this "handle->pending" load can be
reordered before a store that occurs in the client application causing
the write to be missed on the consuming side of uv_async_send(). From the
Intel manual: "Loads May Be Reordered with Earlier Stores to Different
Locations". This is also likely to cause a similar issue on other
architectures that are not as strictly ordered as x86.