unix: Implement `uv_network_interfaces` on Solaris.
The current implementation of `uv_interface_addresses` only works on
Solaris/X86_64 due to relying on `getifaddrs`, which is not provided
on Solaris/SPARC.

This patch implementes the new `uv_network_interfaces` function using
`getifaddrs` on Solaris/X86_64 and using the socket `ioctl` interface
on Solaris/SPARC to retrieve the same information.  Since this new
function provides a superset of the information provided by
`uv_interface_addresses`, the old function is implemented in terms of
the new one and will eventually be removed.