Support uv_backend_fd() on Windows
This commit adds loop option UV_LOOP_EMBED which enables uv_backend_fd()
on Windows for a small performance hit. This is useful when the uv event
loop must be embedded in another event loop. On Windows uv_backend_fd()
will return a handle to a Windows event object which will become signaled
when there are new asynchronous requests to process. When the event is
signaled, uv_run() is guaranteed not to wait. The event may also be set
to signaled by the user to wake up a thread waiting in uv_run() spuriously.

In this mode, the IOCP polling is moved to another thread which signals the
event object when completion port packets have been dequeued. uv_run() polls
by waiting on the event object instead of directly on the completion port
queue.

NOTE: This changes the return type of uv_backend_fd() on Windows from int
to HANDLE.