win, pipe: null-initialize stream->shutdown_req
When initializing a stream, this also inits the shutdown_req field,
instead of waiting for a successfully connection.
A NULL value is used as a sentinel to check for whether this handle is currently
in the shutdown state, and it may not get set if a stream was not connected immediately.