unix: check before reopening tty
uv_tty_init() now checks if the given fd has the same inode as stdin, stdout, or
stderr. Only then the fd is reopened.

This change fixes the wrong assumption that any TTY is bound to /dev/tty. This
is not always the case; file descriptors returned by openpty(3) are independant
from the current TTY.