QUEUE: introduce QUEUE_FOREACH_SAFE() and QUEUE_REMOVE_SAFE()
These macros/functions should be used to remove elements from a queue
while iterating over it.
This turns out to be a common use-case for QUEUE.
For example, calling uv_close() on a handle from the handle's callback
ends up calling QUEUE_REMOVE() from within QUEUE_FOREACH(),
which is unsafe if special care is not taken.

Win backend open-codes a similar thing in src/win/loop-watcher.c, for example.

PR-URL: https://github.com/libuv/libuv/pull/620