unix: make work with Solaris Studio.
The atomics that are in place of a lack of x86 cmpxchg are a GCC
specific intrinsic.  When compiling for Solaris on a SPARC platform
with a non-gcc compiler, this prevents a successful build.  This commit
will rely on a Solaris Studio specific preprocessor macro to redefine
the GCC intrinsic name to be the Solaris Studio one instead.