tests: make fs watch_dir tests more resilient
ReadDirectoryChangesW on Windows can result in multiple invocations of
fs_event_cb for single filesystem events when watching a directory
recursively.  This was resulting in fs_event_watch_dir_recursive
failing.  Upon review, I noticed that the watch_dir tests were written
with the expectation that one create/remove request would result in one
fs_event_cb invocation and that is not stable.

This commit updates the logic that terminates the uv_loop for these
tests to ensure the loop ends only when all creates, all deletes and all
expected fs_event_cb invocations are handled instead of just performing
the latter.  We also remove unnecessary checks to the number of times
fs_event_cb was called since the logic to close the handle already
checkes for this.  This should result in less flakiness in these tests.

Fixes #587