util: extends uv_getrusage by majflt on NT and other NT counters
This extension for the win part of uv_getrusage() adds the
win equivalent of hard page faults to the available properties.

Since there are also similar counters in GetPerformanceInfo() that
are not as easily accessible in posix sytems, it is supposedly
beneficial to add them to  the overall API, while overcoming unices
shortcomings. Note: getrusage() implementation is very different
on unices also.

In consequence uv_rusage_t increaes in size.