src: adding uv_idle_start_nop function
The motivation for this change came from a usecase in Node.js [1]
and the handling of setImmediate functions. These are implemented with
uv_check callbacks. Since checks are performed after polling for I/O,
if there are no idle handler or prepare handler (need to check this) then
the I/O polling would block as there would be nothing for the event loop
to process until there is an I/O event. But if there is an idle handler,
there is something for the event loop to do which will cause the poll
timeout to be zero and the event loop will not block on I/O.

The suggestion here is to add a new function that only takes a
uv_idle_t* and no callback. I first looked at adding using a no-operation
function if the passed in callback was null, but was concerned with
breaking existing functionality as currently if the callback is null
ENIVAL is returned by src/unix/loop-watcher.c and src/win/loop-watcher.c.

I'm was not sure if it would make sense to have no-operation callbacks
for prepare and check handlers too, but if it does the implementations of
these callbacks could probably be moved into loop-watcher.c.

Opening this pull request to get some feedback on what would be the
best way to proceed. Sorry about the long description, but wanted to
verify that my understanding is correct and nothing else.

[1]
https://github.com/nodejs/node/blob/db21266427c1228065dd0576cf59e47ce4485f4e/src/node.cc#L234-L237