CMake build
Hand-written port of the libuv build to CMake. While gyp does provide a
generator for this, it produces a CMake build that is particularly
difficult to work with across the different CMake generators.

I've gone to extra lengths to ensure that the build as written is as
close as possible (minus a tiny delta) to the original gyp build.
Primarily, I've omitted the visual studio / xcode configuration
parameters as CMake sets these appropriately by default. CMake is
capable of handling this with a bit of work, so this could be re-added
later.

Additionally, this hand-written config has been adapted to support
in-tree building and discovery, so that projects can include libuv in
their own projects without installing the library (for local / custom
builds). Just point find_package to the `build` directory and
all of the libuv-* targets will be available.

An install / uninstall target is provided that also provides proper
cmake package configuration definitions, so that find_package will work
with an installed copy.

All required libraries have been defined using find_package, to allow
end users to override locations at configuration time. Use `ccmake` or
the cmake gui for an easy way to see what is needed.

Tests and benchmarks have been ported, but they're currently failing due
to my local machine configuration. They're disabled by default; you can
turn them back on by setting UV_TESTS_ENABLED. YMMV.