Build output modifications
Unifying the build output so the in-tree build is exactly identical to
the installed artifacts on each platform. This nicely simplifies the
overall structure of install as well.

Removing the install generation flag. There isn't any reason install
shouldn't always be generated.

Includes a few format edits as well, nothing major.