build: use '${prefix}' for pkg-config 'exec_prefix'
This follows common pkg-config configuration files and allows for certain
pkg-config overrides to work. For example, when cross-compiling it is common to
have to set a different pkg-config prefix through either '--define-prefix' or a
'PKG_CONFIG_LIBUV_PREFIX' environment variable. If the 'exec_prefix' is
hardcoded to '@prefix@' then it will not pick up those pkg-config overrides.

PR-URL: https://github.com/libuv/libuv/pull/993
Reviewed-By: Ben Noordhuis <info@bnoordhuis.nl>
Reviewed-By: Saúl Ibarra Corretgé <saghul@gmail.com>