win: return IOCP HANDLE in uv_backend_fd
This is a back-port of commit daf8a45d4 from the master branch that
changes the return type of `uv_backend_fd()` from `int` to `uv_os_fd_t`.

It's a non-functional change on Unices because `uv_os_fd_t` is an int.

It should be safe in the API/ABI compatibility sense on Windows because
`uv_backend_fd()` on that platform was previously unusable and therefore
not expected to be in use.

Original commit log follows:

Change the return type from int to uv_os_fd_t so we can return an int on
Unix and a HANDLE on Windows.

Applications can use this to poll the libuv event loop from another
thread as is now possible on Unix.

The test implementation is based on what Electron does:
https://github.com/electron/electron/blob/master/atom/common/node_bindings_win.cc

Refs: https://github.com/libuv/libuv/issues/965
PR-URL: https://github.com/libuv/libuv/pull/1007
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: Imran Iqbal <imran@imraniqbal.org>